Fixed directory placement of outcome classes. Renamed DummyOutcome ---> DefaultOutcome.
