+++ include/aikido/constraint/CollisionFreeOutcome.hpp
